Potential fix for code scanning alert no. 49: Multiplication result converted to larger type
To fix the problem, we need to ensure that the multiplication is performed using the larger integer type (uint64_t) to prevent overflow. This can be achieved by casting one of the operands to uint64_t before performing the multiplication. This way, the multiplication will be done using uint64_t, and the result will be correctly assigned to the uint64_t variable.

+++ libwebp-1.0.3/src/enc/vp8l_enc.c
@@ -1202,7 +1202,7 @@ static void ClearTransformBuffer(VP8LEncoder* const enc) {
 static WebPEncodingError AllocateTransformBuffer(VP8LEncoder* const enc,
                                                  int width, int height) {
   WebPEncodingError err = VP8_ENC_OK;
-  const uint64_t image_size = width * height;
+  const uint64_t image_size = (uint64_t)width * height;
   // VP8LResidualImage needs room for 2 scanlines of uint32 pixels with an extra
   // pixel in each, plus 2 regular scanlines of bytes.
   // TODO(skal): Clean up by using arithmetic in bytes instead of words.
